adding use cases from ICEDIG CDD work
The ICEDIG group working on requirements for a Collections Digitization Dashboard for DiSSCo  - collected use cases. These use cases are structured in a format that provides elements: User group, As a (role), I want to do (x) , so that (y), so I will need (data elements), what is the level of digitization required (specimen, collection, storage, ...), and requires digitization (or not). Adding these use cases for development of the CD standard. Thank you @nielsraes and #icedig